fixed leak in hatch_ant function
