+++ Makefile
@@ -2,8 +2,10 @@ FLAGS=-g
 
 SRC_FILES = $(shell find src -name "*.c")
 INCLUDES = -I./includes -I./libft
-LIBRARY = ./libft/libft.a
+LIBRARY = -L./libft -lft
 RPAD = 45
+
+all: -lemin
 
 -lemin: ./libft/libft.a
 	@printf "%-$(RPAD)s" "Compiling lemin executable. "
@@ -16,3 +18,8 @@ RPAD = 45
 	@$(MAKE) -s -C libft clean
 	@printf "\x1B[32mDone!\x1B[37m\n"
 
+clean:
+	@$(RM) lemin
+
+fclean: clean
+	@$(RM) ./libft/libft.a
