+++ src/pathing_main.c
@@ -1,46 +1,32 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   algo_path.c                                        :+:      :+:    :+:   */
+/*   algo_path_more.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ayano <ayano@student.42.fr>                +#+  +:+       +#+        */
+/*   By: ayano <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2019/09/09 16:16:05 by ayano             #+#    #+#             */
-/*   Updated: 2019/11/05 09:31:03 by ayano            ###   ########.fr       */
+/*   Created: 2019/11/05 09:24:31 by ayano             #+#    #+#             */
+/*   Updated: 2019/11/05 09:28:33 by ayano            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <lem_in.h>
 
-/*
-** Works from end and adds nodes to the struct pathlist.
-*/
-
-t_path			*map_path(t_room *end)
+void		prepend_to_pathlist(t_pathlist **start, t_pathlist *new)
 {
-	t_path *path;
-	t_room *check;
+	t_pathlist *list;
 
-	check = end;
-	while (check->parent)
-		check = check->parent;
-	if (!check->is_start)
-		return (NULL);
-	path = NULL;
-	append_to_path(&path, make_path_item(end));
-	while (end->parent)
+	list = *start;
+	if (!list)
+		*start = new;
+	else
 	{
-		append_to_path(&path, make_path_item(end->parent));
-		end = end->parent;
+		new->next = list;
+		*start = new;
 	}
-	return (path);
 }
 
-/*
-** creates a new node to the path_start linked list.
-*/
-
-t_pathlist		*create_pathlist_item(t_path *path_start)
+t_pathlist	*create_pathlist_item(t_path *path_start)
 {
 	t_pathlist *new;
 
@@ -58,8 +44,7 @@ t_pathlist		*create_pathlist_item(t_path *path_start)
 /*
 ** Mallocs a path node and sets the paths room to the room given as an input.
 */
-
-t_path			*make_path_item(t_room *room)
+t_path		*make_path_item(t_room *room)
 {
 	t_path *new;
 
@@ -75,8 +60,7 @@ t_path			*make_path_item(t_room *room)
 /*
 ** adds the path node called item to end of the path linked list given as start.
 */
-
-void			append_to_path(t_path **start, t_path *item)
+void		append_to_path(t_path **start, t_path *item)
 {
 	t_path	*path;
 
@@ -95,8 +79,7 @@ void			append_to_path(t_path **start, t_path *item)
 /*
 ** adds the node passed as item to the from of the linked list passed as start.
 */
-
-void			prepend_to_path(t_path **start, t_path *item)
+void		prepend_to_path(t_path **start, t_path *item)
 {
 	t_path *path;
