+++ src/stdin_data_functions.c
