+++ src/stdin_data_init.c
