--- visualiser/ant_hatchery.c.bak
@@ -1,34 +1,0 @@
-#include "visualiser.h"
-
-void	join_colony(t_anthill **anthill, t_ant *larvae)
-{
-	t_ant	*nurse;
-	if (!(*anthill)->colony)
-		(*anthill)->colony = larvae;
-	else
-	{
-		nurse = (*anthill)->colony;
-		while (nurse->next)
-			nurse = nurse->next;
-		nurse->next = larvae;
-	}
-}
-
-int		hatch_ant(t_anthill *anthill, int x, int y, char *name)
-{
-	t_ant *egg;
-
-	if (!(egg = malloc(sizeof(t_ant))))
-		return (0);
-	egg->x = x;
-	egg->y = y;
-	egg->is_moving = false;
-	egg->path = NULL;
-	egg->next = NULL;
-	egg->sprite = NULL;
-	egg->following = NULL;
-	egg->current = anthill->start;
-	egg->name = ft_strdup(name);
-	join_colony(&anthill, egg);
-	return (1);
-}
